Dataset
This is a retrospective, multi-center study with data from emergency departments (EDs) from twenty hospitals in different locations in developed countries (Netherlands, USA, and Australia) and developing countries (Botswana, a country in the center of Southern Africa, Pakistan, and Brazil). The sample used in this study includes a selection of twenty EDs. Time series of daily patient arrivals at EDs are from January 1, 2013, to December 30, 2017.